+++ src/color/color.rs
@@ -1,7 +1,8 @@
 use core::fmt;
-use std::collections::HashMap;
 
 use crate::utils;
+
+use super::observer::Observer;
 
 /// Position of the alpha value for hex strings.
 ///
@@ -20,8 +21,8 @@ pub enum AlphaPosition {
 
 //Convert from U32. Needed for converting from the settings AdwComboRow, which use indexes for values.
 impl From<u32> for AlphaPosition {
-    fn from(u: u32) -> Self {
-        match u {
+    fn from(value: u32) -> Self {
+        match value {
             0 => Self::None,
             1 => Self::End,
             2 => Self::Start,
@@ -63,7 +64,7 @@ impl Color {
 
     /// Generate a random color.
     ///
-    /// Although teh RGB values will be randomized, the alpha value will be maximized,
+    /// Although the RGB values will be randomized, the alpha value will be maximized,
     /// so the color will not be transparent.
     pub fn random() -> Self {
         Color::rgb(
@@ -304,13 +305,17 @@ impl Color {
 
     /// Return the colors as CIELAB vales.
     ///
+    /// If use_ten_degrees is true, the function will use 10° observer values instead of the 2° ones.
+    ///
     /// The color will be first converted to XYZ values and then to CIELAB values.
     /// Formula from <http://www.easyrgb.com/en/math.php>
-    pub fn to_cie_lab(self) -> (f32, f32, f32) {
+    pub fn to_cie_lab(self, observer: Observer, use_ten_degrees: bool) -> (f32, f32, f32) {
         //reference xyz for D65 (sRGB) from http://www.easyrgb.com/en/math.php
-        let reference_x = 95.047;
-        let reference_y = 100.000;
-        let reference_z = 108.883;
+        let (reference_x, reference_y, reference_z) = if use_ten_degrees {
+            observer.ten_degrees()
+        } else {
+            observer.two_degrees()
+        };
 
         let xyz = self.to_xyz();
 
@@ -343,9 +348,13 @@ impl Color {
         (cie_l, cie_a, cie_b)
     }
 
-    pub fn to_hcl(self) -> (f32, f32, f32) {
+    /// Convert the color to hcl/ CIELCh
+    ///
+    /// This steps involves converting the color to CIElab first.
+    /// If use_ten_degrees is true, the function will use 10° observer values instead of the 2° ones.
+    pub fn to_hcl(self, observer: Observer, use_ten_degree: bool) -> (f32, f32, f32) {
         //convert color to lab first
-        let (luminance, a, b) = self.to_cie_lab();
+        let (luminance, a, b) = self.to_cie_lab(observer, use_ten_degree);
 
         let hue = b.atan2(a).to_degrees();
         let chroma = (a.powi(2) + b.powi(2)).sqrt();
@@ -440,7 +449,7 @@ impl Color {
                 if t < (2f32 / 3f32) {
                     return p + (q - p) * (2f32 / 3f32 - t) * 6f32;
                 }
-                return p;
+                p
             }
 
             let q = if lightness < 0.5 {