+++ src/widgets/mod.rs
@@ -1,5 +1,4 @@
 pub mod about_window;
-pub mod color_model_entry;
-pub mod hex_entry;
+pub mod color_format_row;
 pub mod palette_dialog;
 pub mod preferences;