+++ CHANGELOG.md
@@ -8,16 +8,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Options to show names of color (from w3c basic, extended and xkcd)
+- Change the color by typing a name in the name field
 - Palettes are now shown in the palette dialog, accessible by clicking on the large color button
 - The App now uses the AdwAboutWindow
 - HSL does now support alpha values
 - HWB and CIELCh are now supported
+- CIE standard observer for CIELab and CIELCh
+-  Change CIE illuminants
 - The following translations have been added
     - Spanish by [IngrownMink4](https://github.com/IngrownMink4)  
     - Basque by [IngrownMink4](https://github.com/IngrownMink4)  
     - Turkish by [libreajans](https://github.com/libreajans)  
     - Dutch by [flipflop97](https://github.com/flipflop97)  
-    - Italian by [phaerrax](https://github.com/phaerrax)  
+    - Italian by [phaerrax](https://github.com/phaerrax)
+- Code refactoring 
 
 ### Removed
 - Removed explicit dependency on the `libc` crate, as it is no longer needed with the new palette dialog